requirements.txt
PyYAML
opencv-python
schedule
pywin32
PyAutoGUI